### Description of the PR
An optional addon pr to PermissionLevelsRemove Client

This, in theory, should make PermissionLevels faster to run, since it doesn't have to get every iteration, and it will skip iterating what was previously empty levels.

This requires heavy testing, and may not make it into the base branch

### Changes Proposed in this Pull Request (List new items in CHANGELOG.MD)

-

### Semver Classification

- [ ] This PR only includes documentation or non-code changes.
- [ ] This PR fixes a bug and does not change the (intended) framework interface.
- [ ] This PR adds methods or properties to the framework interface.
- [x] This PR removes or renames methods or properties in the framework interface.