@bdistin 2018-11-21
initial benchmarks show this may, in fact, be a performance regression... continuing to check more things.