@kyranet 2018-11-21
Besides the O(n) performance requirement to test the last level, benchmarks show a performance decrease by a factor of ~2.1.

<details>
  	<summary>Master benchmark</summary>
	<h2>Run 1</h2>
	<li>0: 252.205ms</li>
	<li>1: 234.281ms</li>
	<li>2: 253.038ms</li>
	<li>3: 254.197ms</li>
	<li>4: 244.717ms</li>
	<li>5: 223.702ms</li>
	<li>6: 271.842ms</li>
	<li>7: 250.372ms</li>
	<li>8: 226.000ms</li>
	<li>9: 251.318ms</li>
	<li>10: 262.683ms</li>
	<h2>Run 2</h2>
	<li>0: 268.072ms</li>
	<li>1: 252.158ms</li>
	<li>2: 270.701ms</li>
	<li>3: 271.369ms</li>
	<li>4: 247.083ms</li>
	<li>5: 254.343ms</li>
	<li>6: 272.429ms</li>
	<li>7: 257.927ms</li>
	<li>8: 262.905ms</li>
	<li>9: 264.659ms</li>
	<li>10: 255.617ms</li>
	<h2>Run 3</h2>
	<li>0: 261.604ms</li>
	<li>1: 269.316ms</li>
	<li>2: 260.825ms</li>
	<li>3: 235.483ms</li>
	<li>4: 256.380ms</li>
	<li>5: 234.020ms</li>
	<li>6: 230.194ms</li>
	<li>7: 246.260ms</li>
	<li>8: 247.823ms</li>
	<li>9: 228.070ms</li>
	<li>10: 258.837ms</li>
	<h2>Run 4</h2>
	<li>0: 232.373ms</li>
	<li>1: 253.942ms</li>
	<li>2: 240.437ms</li>
	<li>3: 227.927ms</li>
	<li>4: 226.728ms</li>
	<li>5: 235.117ms</li>
	<li>6: 218.658ms</li>
	<li>7: 265.600ms</li>
	<li>8: 232.385ms</li>
	<li>9: 221.454ms</li>
	<li>10: 222.144ms</li>
	<h2>Run 5</h2>
	<li>0: 226.984ms</li>
	<li>1: 220.330ms</li>
	<li>2: 269.738ms</li>
	<li>3: 226.516ms</li>
	<li>4: 240.545ms</li>
	<li>5: 225.856ms</li>
	<li>6: 221.963ms</li>
	<li>7: 234.957ms</li>
	<li>8: 219.634ms</li>
	<li>9: 222.948ms</li>
	<li>10: 221.696ms</li>
	<hr>
</details>

<details>
  	<summary>PermissionLevelsRemoveClient benchmark</summary>
	<h2>Run 1</h2>
	<li>0: 286.835ms</li>
	<li>1: 288.363ms</li>
	<li>2: 231.963ms</li>
	<li>3: 215.764ms</li>
	<li>4: 223.675ms</li>
	<li>5: 230.550ms</li>
	<li>6: 230.833ms</li>
	<li>7: 214.579ms</li>
	<li>8: 213.374ms</li>
	<li>9: 220.500ms</li>
	<li>10: 222.476ms</li>
	<h2>Run 2</h2>
	<li>0: 248.225ms</li>
	<li>1: 235.145ms</li>
	<li>2: 214.536ms</li>
	<li>3: 227.082ms</li>
	<li>4: 223.490ms</li>
	<li>5: 218.709ms</li>
	<li>6: 217.661ms</li>
	<li>7: 250.608ms</li>
	<li>8: 217.568ms</li>
	<li>9: 225.734ms</li>
	<li>10: 214.286ms</li>
	<h2>Run 3</h2>
	<li>0: 232.797ms</li>
	<li>1: 313.121ms</li>
	<li>2: 259.910ms</li>
	<li>3: 243.157ms</li>
	<li>4: 236.112ms</li>
	<li>5: 233.343ms</li>
	<li>6: 264.757ms</li>
	<li>7: 255.666ms</li>
	<li>8: 246.976ms</li>
	<li>9: 240.054ms</li>
	<li>10: 250.035ms</li>
	<h2>Run 4</h2>
	<li>0: 336.915ms</li>
	<li>1: 217.423ms</li>
	<li>2: 217.161ms</li>
	<li>3: 219.214ms</li>
	<li>4: 233.910ms</li>
	<li>5: 221.716ms</li>
	<li>6: 217.562ms</li>
	<li>7: 216.575ms</li>
	<li>8: 216.685ms</li>
	<li>9: 215.515ms</li>
	<li>10: 219.185ms</li>
	<h2>Run 5</h2>
	<li>0: 215.065ms</li>
	<li>1: 244.650ms</li>
	<li>2: 212.507ms</li>
	<li>3: 213.182ms</li>
	<li>4: 223.511ms</li>
	<li>5: 218.440ms</li>
	<li>6: 214.777ms</li>
	<li>7: 213.386ms</li>
	<li>8: 214.203ms</li>
	<li>9: 243.549ms</li>
	<li>10: 223.377ms</li>
	<hr>
</details>

<details>
  	<summary>FasterPermissionLevels benchmark</summary>
	<h2>Run 1</h2>
	<li>0: 452.943ms</li>
	<li>1: 468.533ms</li>
	<li>2: 489.328ms</li>
	<li>3: 490.099ms</li>
	<li>4: 510.225ms</li>
	<li>5: 494.746ms</li>
	<li>6: 491.520ms</li>
	<li>7: 530.525ms</li>
	<li>8: 498.952ms</li>
	<li>9: 478.381ms</li>
	<li>10: 486.336ms</li>
	<h2>Run 2</h2>
	<li>0: 474.420ms</li>
	<li>1: 498.887ms</li>
	<li>2: 493.804ms</li>
	<li>3: 520.135ms</li>
	<li>4: 487.868ms</li>
	<li>5: 492.663ms</li>
	<li>6: 698.328ms</li>
	<li>7: 477.055ms</li>
	<li>8: 462.795ms</li>
	<li>9: 597.842ms</li>
	<li>10: 937.933ms</li>
	<h2>Run 3</h2>
	<li>0: 675.565ms</li>
	<li>1: 505.342ms</li>
	<li>2: 476.027ms</li>
	<li>3: 504.224ms</li>
	<li>4: 558.809ms</li>
	<li>5: 501.855ms</li>
	<li>6: 561.021ms</li>
	<li>7: 474.051ms</li>
	<li>8: 483.147ms</li>
	<li>9: 526.850ms</li>
	<li>10: 507.505ms</li>
	<h2>Run 4</h2>
	<li>0: 528.781ms</li>
	<li>1: 520.118ms</li>
	<li>2: 476.594ms</li>
	<li>3: 487.205ms</li>
	<li>4: 510.959ms</li>
	<li>5: 464.877ms</li>
	<li>6: 481.272ms</li>
	<li>7: 437.809ms</li>
	<li>8: 434.005ms</li>
	<li>9: 428.034ms</li>
	<li>10: 407.706ms</li>
	<h2>Run 5</h2>
	<li>0: 404.177ms</li>
	<li>1: 416.782ms</li>
	<li>2: 442.176ms</li>
	<li>3: 414.342ms</li>
	<li>4: 405.285ms</li>
	<li>5: 410.371ms</li>
	<li>6: 448.589ms</li>
	<li>7: 413.968ms</li>
	<li>8: 411.525ms</li>
	<li>9: 443.710ms</li>
	<li>10: 466.619ms</li>
</details>