@kyranet 2018-11-21
### `master`:
```javascript
0 x 1,474,629 ops/sec ±4.44% (61 runs sampled)
1 x 236,512 ops/sec ±3.51% (78 runs sampled)
2 x 233,009 ops/sec ±3.93% (81 runs sampled)
3 x 228,508 ops/sec ±3.78% (80 runs sampled)
4 x 239,684 ops/sec ±3.15% (82 runs sampled)
5 x 247,586 ops/sec ±2.38% (82 runs sampled)
6 x 253,779 ops/sec ±2.45% (86 runs sampled)
7 x 805,839 ops/sec ±3.75% (72 runs sampled)
8 x 1,358,944 ops/sec ±7.33% (37 runs sampled)
9 x 1,415,613 ops/sec ±4.46% (49 runs sampled)
10 x 1,296,332 ops/sec ±4.34% (67 runs sampled)
```

### `PermissionLevelsRemoveClient`:
```javascript
0 x 1,396,291 ops/sec ±4.51% (59 runs sampled)
1 x 218,919 ops/sec ±3.16% (76 runs sampled)
2 x 220,523 ops/sec ±4.70% (78 runs sampled)
3 x 234,376 ops/sec ±3.25% (82 runs sampled)
4 x 231,885 ops/sec ±3.42% (79 runs sampled)
5 x 237,677 ops/sec ±3.20% (82 runs sampled)
6 x 232,729 ops/sec ±2.64% (82 runs sampled)
7 x 825,651 ops/sec ±4.19% (71 runs sampled)
8 x 1,334,120 ops/sec ±4.47% (63 runs sampled)
9 x 1,410,481 ops/sec ±7.30% (45 runs sampled)
10 x 1,442,352 ops/sec ±5.88% (41 runs sampled)
```

### `FasterPermissionLevels`:
```javascript
0 x 951,107 ops/sec ±4.32% (54 runs sampled)
1 x 1,196,449 ops/sec ±3.13% (64 runs sampled)
2 x 844,840 ops/sec ±4.41% (33 runs sampled)
3 x 821,413 ops/sec ±3.82% (32 runs sampled)
4 x 788,730 ops/sec ±4.99% (25 runs sampled)
5 x 857,601 ops/sec ±3.37% (39 runs sampled)
6 x 757,780 ops/sec ±7.56% (24 runs sampled)
7 x 1,143,691 ops/sec ±4.03% (68 runs sampled)
8 x 816,330 ops/sec ±3.54% (24 runs sampled)
9 x 1,152,159 ops/sec ±4.16% (57 runs sampled)
10 x 1,186,724 ops/sec ±2.45% (70 runs sampled)
```

<details>
  	<summary>Benchmark code</summary>
<pre lang="javascript">
const { KlasaClient } = require('klasa');
const client = new KlasaClient();
client.login();
client.on('ready', async () => {
	const guild = client.guilds.get('397700693131264000');
	const channel = guild.channels.get('397701415620837406');
	const message = await channel.messages.fetch('514829238923427844');
	const { permissionLevels } = client;
	// Destroy the client and remove all listeners to avoid fluctuations
	client.destroy();
	client.removeAllListeners();
	new (require('benchmark').Suite)('master')
		.add('0', () => permissionLevels.run(message, 0))
		.add('1', () => permissionLevels.run(message, 1))
		.add('2', () => permissionLevels.run(message, 2))
		.add('3', () => permissionLevels.run(message, 3))
		.add('4', () => permissionLevels.run(message, 4))
		.add('5', () => permissionLevels.run(message, 5))
		.add('6', () => permissionLevels.run(message, 6))
		.add('7', () => permissionLevels.run(message, 7))
		.add('8', () => permissionLevels.run(message, 8))
		.add('9', () => permissionLevels.run(message, 9))
		.add('10', () => permissionLevels.run(message, 10))
		.run({ async: true })
		.on('cycle', (event) => {
			console.log(String(event.target));
		})
		.on('complete', function complete() {
			// eslint-disable-next-line no-invalid-this
			console.log(`Fastest is ${this.filter('fastest').map('name')}`);
		});
});
</pre>
</details>