@bdistin 2018-11-21
I don't like how the code ended up to gain the increased performance, closing this